+++ wikitextparser/_tag.py
@@ -10,7 +10,6 @@ For more info see:
 """
 
 from typing import Dict, Optional, Any
-from warnings import warn
 
 from regex import compile as regex_compile, VERBOSE, DOTALL
 
@@ -19,7 +18,7 @@ from ._wikitext import SubWikiText
 
 # HTML elements all have names that only use alphanumeric ASCII characters
 # https://www.w3.org/TR/html5/syntax.html#syntax-tag-name
-TAG_NAME = rb'(?P<name>[A-Za-z0-9]++)'
+TAG_NAME = rb'(?<name>[A-Za-z0-9]++)'
 # https://www.w3.org/TR/html5/infrastructure.html#space-character
 SPACE_CHARS = rb' \t\n\u000C\r'
 # http://stackoverflow.com/a/93029/2705757
@@ -28,31 +27,29 @@ SPACE_CHARS = rb' \t\n\u000C\r'
 CONTROL_CHARS = rb'\x00-\x1f\x7f-\x9f'
 # https://www.w3.org/TR/html5/syntax.html#syntax-attributes
 ATTR_NAME = (
-    rb'(?P<attr_name>[^' + SPACE_CHARS + CONTROL_CHARS + rb'\u0000"\'>/=]++)'
-)
+    rb'(?<attr_name>[^' + SPACE_CHARS + CONTROL_CHARS + rb'\u0000"\'>/=]++)')
 WS_EQ_WS = rb'[' + SPACE_CHARS + rb']*+=[' + SPACE_CHARS + rb']*+'
 UNQUOTED_ATTR_VAL = (
-    rb'(?P<attr_value>[^' + SPACE_CHARS + rb'"\'=<>`]++)'
-)
-QUOTED_ATTR_VAL = rb'(?P<quote>[\'"])(?P<attr_value>.+?)(?P=quote)'
+    rb'(?<attr_value>[^' + SPACE_CHARS + rb'"\'=<>`]++)')
+QUOTED_ATTR_VAL = rb'(?<quote>[\'"])(?<attr_value>.+?)(?P=quote)'
 # May include character references, but for now, ignore the fact that they
 # cannot contain an ambiguous ampersand.
 ATTR_VAL = (
     # If an empty attribute is to be followed by the optional
     # "/" character, then there must be a space character separating
     # the two. This rule is ignored here.
-    rb'(?:' + WS_EQ_WS + UNQUOTED_ATTR_VAL + rb'[' + SPACE_CHARS + rb']*|'
+    rb'(?:'
+    + WS_EQ_WS + UNQUOTED_ATTR_VAL + rb'[' + SPACE_CHARS + rb']*|'
     + WS_EQ_WS + QUOTED_ATTR_VAL + rb'[' + SPACE_CHARS + rb']*|'
-    + rb'[' + SPACE_CHARS + rb']*+(?P<attr_value>)'  # empty attribute
-    + rb')'
-)
+    + rb'[' + SPACE_CHARS + rb']*+(?<attr_value>)'  # empty attribute
+    + rb')')
 # Ignore ambiguous ampersand for the sake of simplicity.
 ATTR_PATTERN = (
-    rb'(?P<attr>[' + SPACE_CHARS + rb']++' + ATTR_NAME + ATTR_VAL + rb')'
-)
+    rb'(?<attr>[' + SPACE_CHARS + rb']++' + ATTR_NAME + ATTR_VAL + rb')')
 ATTRS_MATCH = regex_compile(
     # Leading space is not required at the start of the attribute string.
-    rb'(?P<attr>[' + SPACE_CHARS + rb']*+' + ATTR_NAME + ATTR_VAL + rb')*+',
+    rb'(?<attr>[' + SPACE_CHARS + rb']*+' + ATTR_NAME + ATTR_VAL + rb')*+'
+    rb'(?<attr_insert>)',
 ).match
 # VOID_ELEMENTS = (
 #     'area', 'base', 'br', 'col', 'embed', 'hr', 'img', 'input', 'keygen',
@@ -64,24 +61,23 @@ ATTRS_MATCH = regex_compile(
 # yet. See
 # https://developer.mozilla.org/en/docs/Web/SVG/Namespaces_Crash_Course
 # for an overview.
-END_TAG_PATTERN = rb'(?P<end></{name}[' + SPACE_CHARS + rb']*+>)'
-END_TAG = END_TAG_PATTERN.replace(rb'{name}', rb'(?P<end_name>(?P=name))')
-TAG_CONTENTS = rb'(?P<contents>.*?)'
+END_TAG_PATTERN = rb'(?<end_tag></{name}[' + SPACE_CHARS + rb']*+>)'
+END_TAG = END_TAG_PATTERN.replace(rb'{name}', rb'(?<end_name>(?P=name))')
+TAG_CONTENTS = rb'(?<contents>.*?)'
 # Note that the following regex won't check for nested tags
 TAG_FULLMATCH = regex_compile(
     rb'''
     # Note that the start group does not include the > character
-    (?P<start>
-        <''' + TAG_NAME + ATTR_PATTERN + rb'''*  # Todo: Possessive?
-    )
+    <''' + TAG_NAME + ATTR_PATTERN + rb'''*  # Todo: Possessive?
     # After the attributes, or after the tag name if there are no attributes,
     # there may be one or more space characters. This is sometimes required but
     # ignored here.
+    (?<attr_insert>)
     [''' + SPACE_CHARS + rb''']*+
     (?>
-        (?P<self_closing>/>)|
+        (?<self_closing>/>)|
         >''' + TAG_CONTENTS + END_TAG + rb'''|
-        (?P<start_only>>)
+        >  # only start; no end tag
     )
     ''',
     DOTALL | VERBOSE,
@@ -89,21 +85,20 @@ TAG_FULLMATCH = regex_compile(
 # Todo: can the tags method be implemented using a TAG_FINDITER? Will
 # that be more performant?
 # TAG_FINDITER should not find any tag containing other tags.
-# TAG_CONTENTS = r'(?P<contents>(?>(?!{TAG}).)*?)'.format(
+# TAG_CONTENTS = r'(?<contents>(?>(?!{TAG}).)*?)'.format(
 #     TAG=TAG.format(**locals())
 # )
 # TAG_FINDITER = regex_compile(
 #     TAG.format(**locals()), flags=DOTALL | VERBOSE
 # ).finditer
 START_TAG_PATTERN = (
-    rb'(?P<start>'
+    rb'(?<start_tag>'
     rb'<{name}(?:' + ATTR_PATTERN + rb')*'
     rb'[' + SPACE_CHARS + rb']*+'
-    rb'(?:(?P<self_closing>/>)|>)'
-    rb')'
-)
+    rb'(?:(?<self_closing>/>)|>)'
+    rb')')
 START_TAG_FINDITER = regex_compile(
-    START_TAG_PATTERN.replace(b'{name}', TAG_NAME)
+    START_TAG_PATTERN.replace(b'{name}', TAG_NAME, 1)
 ).finditer
 
 
@@ -135,11 +130,6 @@ class SubWikiTextWithAttrs(SubWikiText):
             string[s:e] for s, e in self._attrs_match.spans('attr_name')
         )
 
-    def has(self, attr_name: str) -> bool:
-        """Deprecated alias for has_attr."""
-        warn('`has` is deprecated, use `has_attr` instead', DeprecationWarning)
-        return self.has_attr(attr_name)
-
     def get_attr(self, attr_name: str) -> Optional[str]:
         """Return the value of the last attribute with the given name.
 
@@ -155,11 +145,6 @@ class SubWikiTextWithAttrs(SubWikiText):
                 s, e = spans('attr_value')[-i - 1]
                 return string[s:e]
         return None
-
-    def get(self, attr_name: str) -> Optional[str]:
-        """Deprecated alias for get_attr."""
-        warn('`get` is depracated, use `get_attr` instead', DeprecationWarning)
-        return self.get_attr(attr_name)
 
     def set_attr(self, attr_name: str, attr_value: str) -> None:
         """Set the value for the given attribute name.
@@ -179,15 +164,10 @@ class SubWikiTextWithAttrs(SubWikiText):
         # The attr_name is new, add a new attribute.
         fmt = ' {}="{}"' if attr_value else ' {}'
         self.insert(
-            match.end('start'),
+            match.end('attr_insert'),
             fmt.format(attr_name, attr_value)
         )
         return
-
-    def set(self, attr_name: str, attr_value: str) -> None:
-        """Deprecated alias for set_attr."""
-        warn('`set` is depracated, use `set_attr` instead', DeprecationWarning)
-        self.set_attr(attr_name, attr_value)
 
     def del_attr(self, attr_name: str) -> None:
         """Delete all the attributes with the given name.
@@ -202,14 +182,6 @@ class SubWikiTextWithAttrs(SubWikiText):
             if string[s:e] == attr_name:
                 start, stop = match.spans('attr')[-i - 1]
                 del self[start:stop]
-
-    def delete(self, attr_name: str) -> None:
-        """Deprecated alias for del_attr."""
-        warn(
-            '`delete` is depracated, use `del_attr` instead',
-            DeprecationWarning
-        )
-        self.del_attr(attr_name)
 
 
 class Tag(SubWikiTextWithAttrs):